workers/render/requirements.txt
ebooklib
beautifulsoup4
openai
Pillow
piper-tts
pydub
ffmpeg-python
python-dotenv
